requirements.txt
Babel==2.8.0
beautifulsoup4==4.9.3
blinker==1.4
cachelib==0.1.1
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
click==7.1.2
dnspython==2.0.0
email-validator==1.1.1
flake8==3.8.3
Flask==1.1.2
Flask-Assets==2.0
Flask-Babel==2.0.0
Flask-Gravatar==0.5.0
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-Misaka==1.0.0
Flask-Session==0.3.2
Flask-SQLAlchemy==2.4.4
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
jsmin==2.2.2
Markdown==3.3
MarkupSafe==1.1.1
mccabe==0.6.1
misaka==2.1.1
psycopg2==2.8.6
pycodestyle==2.6.0
pycparser==2.20
pyflakes==2.2.0
PyJWT==1.7.1
pyScss==1.3.7
python-dotenv==0.14.0
pytz==2020.1
requests==2.24.0
six==1.15.0
soupsieve==2.0.1
SQLAlchemy==1.3.19
urllib3==1.25.10
webassets==2.0
Werkzeug==1.0.1
